+++ app/src/main/java/com/app/arcabyolimpo/data/repository/supplies/SupplyRepositoryImpl.kt
@@ -93,7 +93,10 @@ class SupplyRepositoryImpl @Inject constructor(
             val result = api.deleteOneSupply(body)
             Log.d("Validacion","Si llego ")
         }
-
+        /** -------------------------------------------------------------------------------------- *
+         * getWorkshopCategoryList -> calls the API to fetch every workshop and category id's and
+         * names.
+         * -------------------------------------------------------------------------------------- */
         override suspend fun getWorkshopCategoryList(): Result<WorkshopCategoryList> {
             return try {
                 val workshopCategoryListDto = api.getWorkshopCategoryList()
@@ -103,7 +106,9 @@ class SupplyRepositoryImpl @Inject constructor(
                 Result.failure(e)
             }
         }
-
+        /** -------------------------------------------------------------------------------------- *
+         * addSupply -> calls the API to add a new Supply and handles the image to rename it
+         * -------------------------------------------------------------------------------------- */
         override suspend fun addSupply(supply: SupplyAdd, image: Uri?): Result<Unit> {
             return try {
                 val imagePart = image?.let {
@@ -116,7 +121,7 @@ class SupplyRepositoryImpl @Inject constructor(
                             context.contentResolver.getType(it)?.toMediaTypeOrNull()
                         )
                         MultipartBody.Part.createFormData(
-                            "SupplyImage",
+                            "imagenInsumo",
                             "image.jpg",
                             requestFile
                         )
@@ -136,7 +141,7 @@ class SupplyRepositoryImpl @Inject constructor(
                     measureUnit = measureUnit,
                     idCategory = idCategory,
                     status = status,
-                    image = imagePart
+                    imagenInsumo = imagePart
                 )
                 Result.success(Unit)
             } catch (e: Exception) {