+++ app/src/main/java/com/app/arcabyolimpo/di/AppModule.kt
@@ -96,7 +96,6 @@ object AppModule {
         authPreferences: UserPreferences,
     ): UserRepository = UserRepositoryImpl(api, authPreferences)
 
-
     @Provides
     @Singleton
     fun providePasswordUserRepository(api: ArcaApi): PasswordUserRepository = PasswordUserRepositoryImpl(api)
@@ -116,7 +115,7 @@ object AppModule {
     @Singleton
     fun provideSupplyRepository(
         api: ArcaApi,
-        @ApplicationContext context: Context
+        @ApplicationContext context: Context,
     ): SupplyRepository = SupplyRepositoryImpl(api, context)
 
     /**